pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Learning Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "awsiotsdk>=1.24.0",
    "fastapi[standard]>=0.115.14",
    "httpx>=0.28.1",
    "pykx>=3.1.4",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]
